create modular structure of WorkspaceFunc component
