+++ src/components/Workspace/TrashcanButtons.js
